imrpove add task menu
